requirements.txt
flask
redis
pymongo
numpy
sentence-transformers
faiss-cpu
requests
